refactor: extract WeeklyEmotionSummary to Components/
Preparatory move so JournalView can render top-5 weekly emotions.
No behavior change.
